+++ src/main/java/com/voodoodyne/trivet/TrivetServlet.java
@@ -1,17 +1,24 @@
 package com.voodoodyne.trivet;
 
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServlet;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
+import java.io.IOException;
 import java.io.Serial;
 import java.util.function.Function;
 
 /**
  * Simple contract that takes the instance mapper function as a constructor parameter.
- * Generally the most convenient way to create the invoker servlet.
+ * Generally the most convenient way to create the invoker servlet. You can also use
+ * the TrivetServer directly in something like Spring.
  */
-public class TrivetServlet extends AbstractTrivetServlet {
+public class TrivetServlet extends HttpServlet {
 	@Serial
 	private static final long serialVersionUID = 1L;
 
-	private final Function<Class<?>, Object> instanceMapper;
+	private final TrivetServer trivetServer;
 
 	/**
 	 * @param instanceMapper should be something like guice's Injector::getInstance or
@@ -20,11 +27,23 @@ public class TrivetServlet extends AbstractTrivetServlet {
 	 *                       Write it by hand if you like.
 	 */
 	public TrivetServlet(final Function<Class<?>, Object> instanceMapper) {
-		this.instanceMapper = instanceMapper;
+		this(new TrivetServer(instanceMapper));
+	}
+
+	/**
+	 * In case you want to customize the TrivetServer.
+	 */
+	public TrivetServlet(final TrivetServer trivetServer) {
+		this.trivetServer = trivetServer;
 	}
 
 	@Override
-	public Object getInstance(Class<?> iface) {
-		return instanceMapper.apply(iface);
+	protected void doPost(final HttpServletRequest req, final HttpServletResponse resp) throws ServletException, IOException {
+		if (!Client.APPLICATION_JAVA_SERIALIZED_OBJECT.equals(req.getContentType()))
+			throw new ServletException("Content-Type must be " + Client.APPLICATION_JAVA_SERIALIZED_OBJECT);
+
+		resp.setContentType(Client.APPLICATION_JAVA_SERIALIZED_OBJECT);
+
+		trivetServer.execute(req.getInputStream(), resp.getOutputStream());
 	}
 }